requirements.txt
# Core dependencies for Kindle to Text Converter

# Screenshot and Window Management
pyautogui==0.9.54
mss==9.0.1
pygetwindow==0.0.9
pywinauto==0.6.8

# Image Processing
Pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4

# OCR
yomitoku>=0.1.0
pytesseract==0.3.13

# Configuration and Data Management
PyYAML==6.0.2
python-dotenv==1.0.1

# CLI and UI
click==8.1.7
tqdm==4.66.5
colorama==0.4.6

# Utilities
loguru==0.7.2
python-dateutil==2.9.0.post0

# Note: If yomitoku installation fails via pip, install manually following:
# https://kotaro-kinoshita.github.io/yomitoku/installation/
